refactor: split streaming related hooks
