+++ packages/next/client/streaming/index.ts
@@ -1,3 +1,3 @@
-export { useRefreshRoot as unstable_useRefreshRoot } from './refresh'
-export { useWebVitalsReport as unstable_useWebVitalsReport } from './vitals'
+export { useRefreshRoot as unstable_useRefreshRoot } from './use-refresh-root'
+export { useWebVitalsReport as unstable_useWebVitalsReport } from './use-webvitals-report'
 export { useFlushEffects as unstable_useFlushEffects } from '../../shared/lib/flush-effects'
