+++ packages/next/client/streaming/use-refresh-root.ts
@@ -1,0 +1,6 @@
+import { useContext } from 'react'
+import { RefreshContext } from './refresh'
+
+export function useRefreshRoot() {
+  return useContext(RefreshContext)
+}
